+++ test/sql/functions.sql
@@ -37,6 +37,12 @@ SELECT l1_distance('[0,0]', '[0,1]');
 SELECT l1_distance('[1,2]', '[3]');
 SELECT l1_distance('[3e38]', '[-3e38]');
 
+SELECT hamming_distance('\xFF', '\xFF');
+SELECT hamming_distance('\xFF', '\xFE');
+SELECT hamming_distance('\xFF', '\xFC');
+SELECT hamming_distance('\xFF', '\x00');
+SELECT hamming_distance('\xFF', '\x0000');
+
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]']) v;
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]', NULL]) v;
 SELECT avg(v) FROM unnest(ARRAY[]::vector[]) v;